stm401: Support for uncalibrated sensors
Added support for uncalibrated gyro and mag sensors.
Modified non wakeable config and status to 3 bytes

Reviewed-on: http://gerrit.pcs.mot.com/616279
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/622591
SLTApproved: Andrew Wheeler <awheeler@motorola.com>
Submit-Approved: Andrew Wheeler <awheeler@motorola.com>
Tested-by: Andrew Wheeler <awheeler@motorola.com>

+++ drivers/misc/stm401/stm401_core.c
@@ -69,7 +69,7 @@ unsigned short stm401_g_mag_delay;
 unsigned short stm401_g_gyro_delay;
 unsigned short stm401_g_baro_delay;
 unsigned short stm401_g_step_counter_delay;
-unsigned short stm401_g_nonwake_sensor_state;
+unsigned long stm401_g_nonwake_sensor_state;
 unsigned short stm401_g_wake_sensor_state;
 unsigned short stm401_g_algo_state;
 unsigned char stm401_g_motion_dur;
